Some tinkering with the rules
This started out just wanting to make it clear that the code of
conduct applied online. This resulted in some editing of the openin
paragraph, which resulted in some editing of where in the list things
went, which resulted in me splitting and elaborating on some of the
items.

I think I've largely preserved the original intent. Please tell me if
you think I've screwed up something important.