Proposing next Austin meeting
Austin folks - we can use this issue to discuss and finalise date and location for the next meet-up. I've done the simplest thing possible by simply saying the same venue, two weeks later. Please discuss!

Once there's more of a consensus, we can get this merged to make the decision concrete. This will have the neat side effect of also updating the site ;-)